Bigger update! Shrunk the world. Preparation to collision render.

+++ WogView/ViewWindow.OnResize.cs
@@ -10,9 +10,9 @@ public partial class ViewWindow
         Console.WriteLine($"Resized from:{ClientSize}, to:{e.Size}");
         int w = e.Width, h = e.Height;
         if (w > h)
-            _camera.AspectRatio = w / h;
+            _camera.AspectRatio = w / (float)h;
         else
-            _camera.AspectRatio = h / w;
+            _camera.AspectRatio = h / (float)w;
         GL.Viewport(0, 0, w, h);
     }
 }
